@@ -1,6 +1,6 @@
 {
   "name": "@seneca/gateway-express",
-  "version": "0.1.1",
+  "version": "0.8.2",
   "main": "dist/gateway-express.js",
   "type": "commonjs",
   "types": "dist/gateway-express.d.ts",
@@ -23,10 +23,10 @@
     "watch": "tsc -w -d",
     "build": "tsc -d",
     "clean": "rm -rf node_modules yarn.lock package-lock.json",
-    "reset": "npm run clean && npm i && npm test",
+    "reset": "npm run clean && npm i && npm run build && npm test",
     "repo-tag": "REPO_VERSION=`node -e \"console.log(require('./package').version)\"` && echo TAG: v$REPO_VERSION && git commit -a -m v$REPO_VERSION && git push && git tag v$REPO_VERSION && git push --tags;",
     "repo-publish": "npm run clean && npm i && npm run repo-publish-quick",
-    "repo-publish-quick": "npm run build && npm run test && npm run repo-tag && npm publish --access public --registry http://registry.npmjs.org "
+    "repo-publish-quick": "npm run build && npm run test && npm run repo-tag && npm publish --access public --registry https://registry.npmjs.org "
   },
   "license": "MIT",
   "files": [
@@ -34,23 +34,23 @@
     "src/**/*.ts",
     "LICENSE"
   ],
+  "engines": {
+    "node": ">=12"
+  },
   "devDependencies": {
-    "@seneca/gateway": "^0.1.5",
-    "@seneca/maintain": "^0.0.19",
-    "@types/jest": "^27.0.1",
-    "esbuild": "^0.12.25",
+    "@seneca/maintain": "^0.0.22",
+    "@types/jest": "^28.1.6",
+    "esbuild": "^0.15.1",
     "esbuild-jest": "^0.5.0",
     "express": "^4.17.1",
-    "jest": "^27.1.0",
-    "seneca": "^3.23.3",
-    "seneca-msg-test": "^3.2.0",
-    "seneca-promisify": "^2.0.0",
-    "typescript": "^4.4.2"
+    "jest": "^28.1.3",
+    "seneca-msg-test": "^3.4.0",
+    "typescript": "^4.7.4"
   },
   "peerDependencies": {
-    "@seneca/gateway": ">=0",
+    "@seneca/gateway": ">=0.5.1",
     "express": ">=4",
     "seneca": ">=3",
-    "seneca-promisify": ">=2"
+    "seneca-promisify": ">=3"
   }
 }
